+++ packages/components/card/src/AssetCard/AssetCard.styles.ts
@@ -1,5 +1,5 @@
-import { css } from 'emotion';
 import tokens from '@contentful/f36-tokens';
+import { css } from 'emotion';
 
 export const getAssetCardStyles = () => {
   return {