+++ packages/components/datetime/src/types.ts
@@ -1,3 +1,8 @@
 export type DateType = Date | string | number;
 
-export type DateFormat = 'full' | 'time' | 'weekday' | 'day';
+export type DateFormat =
+  | 'fullWithSeconds'
+  | 'full'
+  | 'time'
+  | 'day'
+  | 'weekday';