+++ packages/components/datetime/src/utils/formatDateTimeUtils.ts
@@ -35,6 +35,9 @@ export function formatDateAndTime(
     case 'time':
       template = 'h:mm A'; // 3:45 PM
       break;
+    case 'fullWithSeconds':
+      template = 'ddd, DD MMM YYYY [at] h:mm:ss A'; // Tue, 17 Aug 2021 at 3:45:67 PM
+      break;
     default:
       template = 'ddd, DD MMM YYYY [at] h:mm A'; // Tue, 17 Aug 2021 at 3:45 PM
   }