+++ packages/website/components/LiveEditor/StaticSource.tsx
@@ -10,6 +10,8 @@ import tokens from '@contentful/f36-tokens';
 const styles = {
   root: css({
     padding: tokens.spacingM,
+    borderRadius: tokens.borderRadiusMedium,
+    marginBottom: tokens.spacingM,
   }),
 };