+++ packages/forma-36-react-components/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.3](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.2...@contentful/f36-components@4.0.3) (2022-01-12)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.2](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.1...@contentful/f36-components@4.0.2) (2022-01-12)
+
+
+### Bug Fixes
+
+* Update README ([#1745](https://github.com/contentful/forma-36/issues/1745)) ([1094eb1](https://github.com/contentful/forma-36/commit/1094eb196d14c008433e1fbdedf6ec11d8570a91))
+
+
+
+
+
+## [4.0.1](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.0...@contentful/f36-components@4.0.1) (2022-01-11)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
 ## [3.100.4](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-components@3.100.3...@contentful/forma-36-react-components@3.100.4) (2021-11-22)
 
 **Note:** Version bump only for package @contentful/forma-36-react-components